requirements.txt
livekit>=0.19.1
livekit-agents>=0.12.11
livekit-plugins-openai>=0.10.17
livekit-plugins-google>=0.3.0
python-dotenv~=1.0
